add endpoint for admin getting seminar data

+++ database/migrations/2021_01_06_105645_drop_column_pembimbing_tiga_on_table_seminars.php
@@ -14,7 +14,11 @@ class DropColumnPembimbingTigaOnTableSeminars extends Migration
     public function up()
     {
         Schema::table('seminars', function($table){
-            $table->dropColumn('pembimbing_tiga');
+            
+            if (Schema::hasColumn('seminars','pembimbing_tiga')){
+                //
+                $table->dropColumn('pembimbing_tiga');    
+            }
         });
     }
